M8_Projects/cub-modelapi/requirements.txt
amqp==2.6.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.2
Django==4.0.3
djangorestframework==3.13.1
greenlet==1.1.2
joblib==1.1.0
kombu==4.6.11
memory-profiler==0.60.0
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.1
psutil==5.9.0
py4j==0.10.9.2
pyspark==3.2.0
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.62.3
vine==1.3.0
xgboost==1.5.1
